requirements.txt
pymupdf
pytesseract
Pillow
openai
python-dotenv
